requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.0
APScheduler==3.6.3
arrow==0.15.4
billiard==3.5.0.2
blinker==1.4
click==6.7
decorator==4.0.11
defusedxml==0.4.1
docutils==0.13.1
filelock==3.0.4
Flask==1.1.1
Flask-Logging-Extras==0.2.0
Flask-WTF==0.14.2
greenlet==0.4.13
ipdb==0.11
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10.3
kombu==4.0.2
MarkupSafe==0.23
packaging==16.8
parso==0.1.1
pbr==1.10.0
pexpect==4.4.0
picamera==1.13
pickleshare==0.7.4
pip-review==0.5.2
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.1.10
pysmbus==0.1
python-dateutil==2.6.0
python3-openid==3.0.10
pytz==2016.10
RPi.GPIO==0.7.0
simplegeneric==0.8.1
six==1.10.0
smbus2==0.2.0
sqlparse==0.2.2
Tempita==0.5.2
traitlets==4.3.2
tzlocal==1.5.1
uWSGI==2.0.17
vine==1.1.3
wcwidth==0.1.7
Werkzeug==0.16.0
WTForms==2.2.1
